+++ packages/eslint-mdx/src/index.ts
@@ -1,3 +1,4 @@
 export * from './helpers'
 export * from './parser'
+export * from './sync'
 export * from './types'
